Create display function for artwork layout
