+++ src/scss/styles.scss
@@ -143,7 +143,7 @@
   gap: 20px;
   padding-block-start: 98px;
 
-  &-item {
+  &__item {
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
